requirements.txt
darkdetect==0.8.0
Pillow==11.0.0
Requests==2.32.3
sv_ttk==2.6.0
notify_py==0.3.43